# Buffer pooling in alert templates