Hi,

I added benchmarks for alert node and found a quick performance boost,
After memory profiling, I found that we are allocating new buffer for the templates of id,message and details, so I wrapped this in a buffer pool, and got a performance improvement:

```
benchmark                            old ns/op       new ns/op       delta
Benchmark_T10_P500_AlertTask-4       210263431       159024737       -24.37%
Benchmark_T10_P50000_AlertTask-4     18718435131     15589249575     -16.72%
Benchmark_T1000_P500_AlertTask-4     21833771619     20820186700     -4.64%

benchmark                            old allocs     new allocs     delta
Benchmark_T10_P500_AlertTask-4       490648         467352         -4.75%
Benchmark_T10_P50000_AlertTask-4     50044467       47560324       -4.96%
Benchmark_T1000_P500_AlertTask-4     48492207       46626729       -3.85%

benchmark                            old bytes      new bytes      delta
Benchmark_T10_P500_AlertTask-4       36155296       33162568       -8.28%
Benchmark_T10_P50000_AlertTask-4     3686308856     3347320896     -9.20%
Benchmark_T1000_P500_AlertTask-4     3604462496     3463221064     -3.92%
```
_Compared to master_