@yosiat 2016-04-04
by the way, I think we can improve further if we will move from text/json to https://github.com/pquerna/ffjson