@nathanielc 2016-04-04
@yosiat Thanks this looks great. 

> by the way, I think we can improve further if we will move from text/json to https://github.com/pquerna/ffjson

I think we will hold off on this one for now since the number of triggered alerts should be orders of magnitude less than the number of points processed. Adding the extra dependency seems over kill at this point.