@yosiat 2016-04-04
@nathanielc I do agree, but when I profiled the benchmarks I found out that text/json is called a lot, when I don't trigger the alert (I am not calling post(), log(), or something)

It might be in the id/message/details rendering, I will check that out.